requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.2.1
django-autoslug==1.9.9
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
pillow==11.2.1
PyJWT==2.9.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
